clases/recursos_clase04/requirements.txt
# Requerimientos para Clase 04 - Pipeline de Desarrollo Geoespacial
# Python 3.9+

# Core geoespacial
geopandas==0.14.1
pandas==2.1.4
numpy==1.26.2
shapely==2.0.2
fiona==1.9.5
pyproj==3.6.1
rtree==1.1.0

# Visualización
folium==0.15.1
matplotlib==3.8.2
plotly==5.18.0
seaborn==0.13.0

# Web frameworks
streamlit==1.29.0
streamlit-folium==0.17.4
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3

# APIs y networking
requests==2.31.0
googlemaps==4.10.0
osmnx==1.9.0
networkx==3.2.1
geopy==2.4.1

# Base de datos
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pymongo==4.6.1

# Procesamiento paralelo
dask[complete]==2023.12.1
dask-geopandas==0.3.1

# Machine Learning
scikit-learn==1.3.2
pykrige==1.7.0

# Utilidades
python-dotenv==1.0.0
tqdm==4.66.1
click==8.1.7
pyyaml==6.0.1
jupyter==1.0.0
ipykernel==6.27.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Documentación
mkdocs==1.5.3
mkdocs-material==9.5.3